Move tcgmsg and gaf2c under armci. armci now supports tcgmsg directly (since it is included in libarmci). Needed to rename tcgmsg header files to avoid conflicts with armci headers. Fix bug where tcgmsg over MPI was calling MP_INIT macro. Replace MAYBE_RESTRICT with just restrict since configure determines whether restrict is supported and will act appropriately. C versions of accumulate and copy routines are always used. Fix bug of missing symbol parmci_msg_group_barrier when compiling for tcgmsg. Comment out a free() in groups code to avoid seg fault. Move some makefile pieces int *.inc files and use include directive.

+++ armci/tcgmsg/ipcv4.0/tcgshmem.h
@@ -13,7 +13,7 @@
   the id associated with the region. The return vaue is a pointer to the
   the region. Any error is a hard fail.
 
-  (char *) CreateSharedRegion((long *) id, (long *) size)
+  (char *) CreateSharedRegion((Integer *) id, (Integer *) size)
 */
 extern char *CreateSharedRegion();
 
@@ -22,9 +22,9 @@ extern char *CreateSharedRegion();
   -1 for failure. id, size, and addr much match exactly those items returned
   from CreateSharedRegion
 
-  long DetachSharedRegion((long) id, (long) size, (char *) addr)
+  Integer DetachSharedRegion((Integer) id, (Integer) size, (char *) addr)
 */
-extern long DetachSharedRegion();
+extern Integer DetachSharedRegion();
 
 
 /*
@@ -32,9 +32,9 @@ extern long DetachSharedRegion();
   to remove it from the system. On the Alliant the shared region disappears
   when the last process dies or detaches. Returns 0 on success, -1 on error.
 
-  long DeleteSharedRegion( (long) id)
+  Integer DeleteSharedRegion( (Integer) id)
 */
-extern long DeleteSharedRegion();
+extern Integer DeleteSharedRegion();
 
 
 /*
@@ -43,6 +43,6 @@ extern long DeleteSharedRegion();
   from CreateSharedRegion (which in turn is the requested size rounded
   up to a multiple of 4096). Any error is a hard fail. 
 
-  (char *) AttachSharedRegion( (long) id, (long) size))
+  (char *) AttachSharedRegion( (Integer) id, (Integer) size))
 */
 extern char *AttachSharedRegion();
